+++ api/src/pdc/services/company/CompanyServiceProvider.ts
@@ -4,11 +4,11 @@ import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
 
 import { FetchCommand } from "@/pdc/services/company/commands/FetchCommand.ts";
-import { binding as fetchBinding } from "@/shared/company/fetch.schema.ts";
-import { binding as findBinding } from "@/shared/company/find.schema.ts";
 import { FetchAction } from "./actions/FetchAction.ts";
 import { FindAction } from "./actions/FindAction.ts";
 import { config } from "./config/index.ts";
+import { binding as fetchBinding } from "./contracts/fetch.schema.ts";
+import { binding as findBinding } from "./contracts/find.schema.ts";
 import { CompanyDataSourceProvider } from "./providers/CompanyDataSourceProvider.ts";
 import { CompanyRepositoryProvider } from "./providers/CompanyRepositoryProvider.ts";
