+++ api/src/pdc/services/company/contracts/fetch.contract.ts
@@ -1,12 +1,12 @@
-import { CompanyInterface } from '../common/interfaces/CompanyInterface2.ts';
+import { CompanyInterface } from "../../../../shared/common/interfaces/CompanyInterface2.ts";
 
 export type ParamsInterface = string;
 
 export type ResultInterface = CompanyInterface;
 
 export const handlerConfig = {
-  service: 'company',
-  method: 'fetch',
+  service: "company",
+  method: "fetch",
 } as const;
 
 export const signature = `${handlerConfig.service}:${handlerConfig.method}` as const;
