+++ api/src/pdc/services/company/contracts/find.contract.ts
@@ -1,4 +1,4 @@
-import { CompanyInterface } from '../common/interfaces/CompanyInterface2.ts';
+import { CompanyInterface } from "../../../../shared/common/interfaces/CompanyInterface2.ts";
 
 export interface ParamsInterface {
   query: {
@@ -11,8 +11,8 @@ export interface ParamsInterface {
 export interface ResultInterface extends CompanyInterface {}
 
 export const handlerConfig = {
-  service: 'company',
-  method: 'find',
+  service: "company",
+  method: "find",
 } as const;
 
 export const signature = `${handlerConfig.service}:${handlerConfig.method}` as const;
